refactor: split routes/read.rs into feature-based submodules
Closes #99

Split the 1679-line read.rs into 9 submodules under routes/read/:
- access.rs: access check helpers
- dto.rs: all DTO types and shared parse helpers
- repositories.rs: list/get repository handlers
- board_projects.rs: list/get board project handlers
- board_runs.rs: list/get board run handlers
- artifacts.rs: list artifacts handler
- viewer_sources.rs: viewer sources handler + helpers
- diff.rs: board run diff handler
- findings.rs: findings handler + helpers

mod.rs re-exports all public items via glob imports,
maintaining all external paths (lib.rs, api_token.rs).

No behavior changes. All tests pass, OpenAPI snapshot unchanged.
