+++ crates/api/src/routes/read/mod.rs
@@ -1,0 +1,19 @@
+mod access;
+mod artifacts;
+mod board_projects;
+mod board_runs;
+mod diff;
+mod dto;
+mod findings;
+mod repositories;
+mod viewer_sources;
+
+pub use access::*;
+pub use artifacts::*;
+pub use board_projects::*;
+pub use board_runs::*;
+pub use diff::*;
+pub use dto::*;
+pub use findings::*;
+pub use repositories::*;
+pub use viewer_sources::*;
